Restore previous test files
